+++ api4jenkins/job.py
@@ -1,17 +1,18 @@
 # encoding: utf-8
-from _ctypes import ArgumentError
-from _functools import partial
-from pathlib import PurePosixPath
 import json
+from pathlib import PurePosixPath
+
+from _functools import partial
 
 from .credential import Credentials
-from .item import snake, Item, append_slash
-from .mix import ConfigrationMix, DescriptionMix, DeletionMix
+from .item import Item, append_slash, snake
+from .mix import (ConfigurationMixIn, DeletionMixIn, DescriptionMixIn,
+                  EnableMixIn)
 from .queue import QueueItem
 from .view import Views
 
 
-class Job(Item, ConfigrationMix, DescriptionMix, DeletionMix):
+class Job(Item, ConfigurationMixIn, DescriptionMixIn, DeletionMixIn):
 
     def move(self, path):
         path = path.strip('/')
@@ -20,12 +21,14 @@ class Job(Item, ConfigrationMix, DescriptionMix, DeletionMix):
         resp = self.handle_req('POST', 'move/move',
                                data=params, allow_redirects=False)
         self.url = resp.headers['Location']
+        return resp
 
     def rename(self, name):
         resp = self.handle_req('POST', 'confirmRename',
                                params={'newName': name},
                                allow_redirects=False)
         self.url = append_slash(resp.headers['Location'])
+        return resp
 
     def duplicate(self, path):
         self.jenkins.create_job(path, self.configure())
@@ -35,15 +38,14 @@ class Job(Item, ConfigrationMix, DescriptionMix, DeletionMix):
         path = PurePosixPath(self.full_name)
         if path.parent.name == '':
             return self.jenkins
-        return Folder(self.jenkins,
-                      self.jenkins._name2url(str(path.parent)))
+        return self.jenkins.get_job(str(path.parent))
 
 
 class Folder(Job):
 
     def create(self, name, xml):
-        self.handle_req('POST', 'createItem', params={'name': name},
-                        headers=self.headers, data=xml)
+        return self.handle_req('POST', 'createItem', params={'name': name},
+                               headers=self.headers, data=xml)
 
     def get(self, name):
         for item in self.api_json(tree='jobs[name,url]')['jobs']:
@@ -69,11 +71,11 @@ class Folder(Job):
 
     def copy(self, src, dest):
         params = {'name': dest, 'mode': 'copy', 'from': src}
-        self.handle_req('POST', 'createItem', params=params,
-                        allow_redirects=False)
+        return self.handle_req('POST', 'createItem', params=params,
+                               allow_redirects=False)
 
     def reload(self):
-        self.handle_req('POST', 'reload')
+        return self.handle_req('POST', 'reload')
 
     @property
     def views(self):
@@ -88,11 +90,18 @@ class Folder(Job):
         yield from self.iter()
 
 
-class WorkflowMultiBranchProject(Folder):
-    pass
+class WorkflowMultiBranchProject(Folder, EnableMixIn):
+
+    def scan(self, delay=0):
+        return self.handle_req('POST', 'build', params={'delay': delay})
+
+    def get_scan_log(self, stream=False):
+        with self.handle_req('GET', 'indexing/consoleText', stream=stream) as resp:
+            for line in resp.iter_lines():
+                yield line
 
 
-class Project(Job):
+class Project(Job, EnableMixIn):
 
     def __init__(self, jenkins, url):
         super().__init__(jenkins, url)
@@ -126,12 +135,6 @@ class Project(Job):
 
     def iter_builds(self):
         yield from self
-
-    def enable(self):
-        return self.handle_req('POST', 'enable')
-
-    def disable(self):
-        return self.handle_req('POST', 'disable')
 
     def set_next_build_number(self, number):
         self.handle_req('POST', 'nextbuildnumber/submit',