+++ tests/unit/test_system.py
@@ -4,7 +4,8 @@ from api4jenkins.item import snake
 
 class TestSystem:
 
-    @pytest.mark.parametrize('action', ['restart', 'safeRestart', 'quietDown', 'cancelQuietDown'])
+    @pytest.mark.parametrize('action', ['restart', 'safeRestart', 'quietDown',
+                                        'cancelQuietDown', 'exit', 'safeExit'])
     def test_enable_disable(self, jenkins, mock_resp, action):
         req_url = f'{jenkins.system.url}{action}'
         mock_resp.add('POST', req_url)